"Turbo"-charge editing the room name
The trick is define two turbo-frame blocks with the same id; for
example, <turbo-frame id="room"> in this commit---one is on the
view room page, and the other on the edit room page.

Turbo looks for that "magic" element name, and when you click
a link inside one, it
  - intercepts the click,
  - fetches the HTML from the anchor href,
  - grabs the matching turbo-frame block in that HTML, and
  - swaps in that HTML to the view page.

The end result is the form to edit the room now appears on
the room detail page.

Note that if the Javascript does not load (or the user has
turned off JavaScript, the app still works and the full edit
room page displays like it did in the previous commit.