"Turbo"-charge adding a new message
Use the src tag on a turbo-frame tag to load the new message form
into the room detail page.

Change the HTTP response of the new message form from 201 to 302,
so the room detail page is reloaded after a message is posted.

For some reason Safari set the media type of the new message form
post to multipart form after the form was embedded in the room
detail page.  So I modified the message controller to accept that
media type as well.