WIP: connect Turbo to web socket
Use JS generated from Rails demo---the web socket support
in Turbo comes from turbo-rails code and not the core
turbo code.

Implement a TextWebScoket handler, and register it on
/cable endpoint.

Result is that we get a socket connection that is
immediately closed.  Safari network tab shows:

              WebSocket Connection Established
  (up-arrow)  Connection Close Frame
              Connection Close Frame

all in the same second.

Note that none of the debug logging in the web socket
handler is output.